Use enum instead of strings.
This changes the type of all REASON_* constants.
One outcome of this is that they no longer have equality with
their old string values e.g. REASON_FOCUS == "focus" >> False

This would also be the case if the enum used string values instead
of int values.

All occurrences of the previous string values were inspected, and with a
reasonable degree of certainty, there seemed not to be used as REASON_*
values.